refactor(frontend): move user config form logic to context
